# sql: add a benchmark for large DROP DATABASE statements