@ajwerner 2022-04-21
```
./dev test pkg/sql --filter -  -v  -- --test_arg=--test.bench=DropLargeDatabase --test_arg=--test.benchmem --test_arg=--test.benchtime=1x --test_arg=--test.count=7 --test_output=streamed  --cache_test_results=no 2>&1 | tee out
benchstat <( cat out| grep declarative=false  | sed 's/,declarative=[a-z]*//g' )   <( cat out| grep declarative=true  | sed 's/,declarative=[a-z]*//g' )
```
```
name                                             old time/op    new time/op    delta
DropLargeDatabase/tables=3,columns=2,depth=2-16    58.6ms ± 1%   210.9ms ± 3%   +259.83%  (p=0.003 n=5+7)
DropLargeDatabase/tables=3,columns=4,depth=2-16    58.9ms ± 4%   370.5ms ±19%   +528.93%  (p=0.001 n=6+7)
DropLargeDatabase/tables=3,columns=2,depth=3-16    71.4ms ±11%   681.1ms ±35%   +853.65%  (p=0.001 n=6+7)
DropLargeDatabase/tables=3,columns=4,depth=3-16    72.5ms ± 4%  1033.7ms ± 5%  +1325.56%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=2,depth=4-16    88.1ms ± 4%  1644.3ms ±12%  +1765.39%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=4,depth=4-16     120ms ± 2%    3340ms ± 8%  +2674.42%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=2,depth=5-16     183ms ± 4%    4787ms ± 5%  +2521.15%  (p=0.001 n=6+7)
DropLargeDatabase/tables=3,columns=4,depth=5-16     616ms ± 5%   10035ms ± 2%  +1528.44%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=2,depth=2-16    70.6ms ± 6%   386.0ms ± 2%   +446.81%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=2-16    73.2ms ± 3%   693.2ms ± 2%   +847.52%  (p=0.001 n=6+7)
DropLargeDatabase/tables=4,columns=2,depth=3-16    92.3ms ± 8%  1409.8ms ± 6%  +1427.63%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=3-16     119ms ± 9%    2829ms ± 1%  +2281.31%  (p=0.001 n=7+6)
DropLargeDatabase/tables=4,columns=2,depth=4-16     248ms ± 4%    6099ms ± 2%  +2363.58%  (p=0.001 n=6+7)
DropLargeDatabase/tables=4,columns=4,depth=4-16     854ms ± 9%   12870ms ± 1%  +1406.36%  (p=0.001 n=7+6)
DropLargeDatabase/tables=4,columns=2,depth=5-16     4.90s ± 6%    27.93s ± 6%   +470.10%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=5-16     35.3s ± 4%     62.1s ± 6%    +76.15%  (p=0.001 n=7+7)

name                                             old alloc/op   new alloc/op   delta
DropLargeDatabase/tables=3,columns=2,depth=2-16    26.7MB ± 1%    35.6MB ± 1%    +33.31%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=4,depth=2-16    29.3MB ± 1%    55.9MB ± 1%    +90.75%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=2,depth=3-16    36.3MB ± 3%    77.1MB ± 0%   +112.35%  (p=0.001 n=7+6)
DropLargeDatabase/tables=3,columns=4,depth=3-16    43.6MB ± 1%   144.1MB ± 0%   +230.44%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=2,depth=4-16    58.1MB ± 1%   219.4MB ± 0%   +277.32%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=4,depth=4-16    89.3MB ± 1%   541.0MB ± 0%   +505.95%  (p=0.003 n=7+5)
DropLargeDatabase/tables=3,columns=2,depth=5-16     123MB ± 1%     679MB ± 0%   +453.32%  (p=0.001 n=7+6)
DropLargeDatabase/tables=3,columns=4,depth=5-16     260MB ± 0%    1679MB ± 0%   +545.28%  (p=0.002 n=6+6)
DropLargeDatabase/tables=4,columns=2,depth=2-16    35.8MB ± 1%    62.0MB ± 1%    +73.02%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=2-16    41.5MB ± 2%   108.3MB ± 0%   +161.09%  (p=0.001 n=7+6)
DropLargeDatabase/tables=4,columns=2,depth=3-16    58.9MB ± 1%   212.3MB ± 0%   +260.23%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=3-16    85.4MB ± 1%   387.2MB ± 0%   +353.43%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=2,depth=4-16     161MB ± 1%     887MB ± 0%   +451.38%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=4-16     363MB ± 0%    2209MB ± 0%   +509.06%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=2,depth=5-16    1.19GB ± 0%    4.77GB ± 1%   +300.81%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=5-16    4.25GB ± 0%    9.77GB ± 1%   +129.78%  (p=0.001 n=6+7)

name                                             old allocs/op  new allocs/op  delta
DropLargeDatabase/tables=3,columns=2,depth=2-16      197k ± 1%      277k ± 0%    +40.92%  (p=0.002 n=6+6)
DropLargeDatabase/tables=3,columns=4,depth=2-16      204k ± 1%      441k ± 0%   +116.34%  (p=0.003 n=7+5)
DropLargeDatabase/tables=3,columns=2,depth=3-16      250k ± 2%      637k ± 0%   +154.96%  (p=0.001 n=7+6)
DropLargeDatabase/tables=3,columns=4,depth=3-16      272k ± 0%     1162k ± 0%   +327.58%  (p=0.001 n=7+6)
DropLargeDatabase/tables=3,columns=2,depth=4-16      333k ± 1%     1701k ± 0%   +411.50%  (p=0.001 n=7+7)
DropLargeDatabase/tables=3,columns=4,depth=4-16      395k ± 1%     3508k ± 0%   +788.83%  (p=0.001 n=7+6)
DropLargeDatabase/tables=3,columns=2,depth=5-16      505k ± 1%     4950k ± 0%   +879.50%  (p=0.002 n=6+6)
DropLargeDatabase/tables=3,columns=4,depth=5-16      680k ± 0%    10242k ± 0%  +1405.92%  (p=0.004 n=5+6)
DropLargeDatabase/tables=4,columns=2,depth=2-16      259k ± 1%      489k ± 0%    +88.68%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=2-16      276k ± 1%      843k ± 0%   +205.29%  (p=0.001 n=7+6)
DropLargeDatabase/tables=4,columns=2,depth=3-16      361k ± 0%     1613k ± 0%   +346.97%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=3-16      420k ± 1%     3047k ± 0%   +625.66%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=2,depth=4-16      606k ± 1%     6134k ± 0%   +912.45%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=4-16      842k ± 1%    12783k ± 0%  +1417.38%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=2,depth=5-16     1.71M ± 1%    25.34M ± 0%  +1382.89%  (p=0.001 n=7+7)
DropLargeDatabase/tables=4,columns=4,depth=5-16     4.00M ± 1%    50.50M ± 0%  +1161.85%  (p=0.002 n=6+6)
```